Linux 2.2.21pre3
o	Fix a case where a non blocking tty write could	(Peter Benie)
	get stuck
o	Fix non blocking midi close on es1370, es1371	(Alan Cox)
	sonicvibes right this time
o	Fix menu/xconfig warnings			(René Scharfe)
o	Fix non blocking midi close on cmpci, cs4281,	(Alan Cox)
	esssolo, trident.
o	Add eepro100VE ident				(Hanno Boeck)
o	Fix DRM oops case				(Herbert Xu)
o	Fix an oops causing datagram AF_UNIX race	(Paul Menage)
o	Support newer geodes using new CPUID properly	(Hiroshi Miura)
o	Fix up RTC build for non pmac ppc boxes		(Tom Rini)
o	Fix MCE address reporting			(Pete Wyckoff)
o	Vibra16 docs update				(Neale Banks)
o	Eicon include file fix				(Herbert Xu)
o	ISDN loop and header fixes			(Kai Germaschewski)
o	Fix eepro100 out of memory during init path	(Neale Banks)
o	Fix BSD partition table handling breakage	(Andries Brouwer)
o	Add WD XD signature to xd driver		(Paul)
o	3Ware driver update				(Adam Radford)
o	S/390 debugging updates				(Carsten Otte)
o	S/390 DASD updates				(Carsten Otte)
o	S/390 CIO updates				(Carsten Otte)
o	Update USB serial, belkin, digi_acceleport,	(Greg Kroah-Hartmann)
	empeg, ftdsio, edgeport, keyspan, mctu232,
	omninet, prolific, visor
o	Cyberjack USB driver				(Matthias Bruestle)
o	USB ir dongle driver				(Greg Kroah-Hartmann)
o	Support very large FAT file systems		(Vijay Kumar)
o	Backport 2.4 modversions build fix		(Mikael Pettersson)
o	Backport 2.4 es1371 init for new revs		(Julian Anastasov)
o	3c507 driver fixes				(Mark Mackenzie)
o	ext2 obscure group descriptor corruption fix	(Daniel Phillips,
							 Al Viro)
o	Correct a problem where rpciod didnt give up	(Andreas Haumer)
	its current dir

+++ fs/fat/misc.c
@@ -349,10 +349,10 @@ void fat_date_unix2dos(int unix_date,unsigned short *time,
  */
 
 int fat__get_entry(struct inode *dir, loff_t *pos,struct buffer_head **bh,
-    struct msdos_dir_entry **de, int *ino)
+    struct msdos_dir_entry **de, loff_t *ino)
 {
 	struct super_block *sb = dir->i_sb;
-	int sector, offset;
+	long long sector, offset;
 
 	while (1) {
 		offset = *pos;
@@ -428,8 +428,8 @@ int fat__get_entry(struct inode *dir, loff_t *pos,struct buffer_head **bh,
 	    (*number)++; \
     }
 
-static int raw_scan_sector(struct super_block *sb,int sector,const char *name,
-    int *number,int *ino,struct buffer_head **res_bh,
+static int raw_scan_sector(struct super_block *sb,long long sector,
+    const char *name, int *number,loff_t *ino,struct buffer_head **res_bh,
     struct msdos_dir_entry **res_de)
 {
 	struct buffer_head *bh;
@@ -475,7 +475,7 @@ static int raw_scan_sector(struct super_block *sb,int sector,const char *name,
  * requested entry is found or the end of the directory is reached.
  */
 
-static int raw_scan_root(struct super_block *sb,const char *name,int *number,int *ino,
+static int raw_scan_root(struct super_block *sb,const char *name,int *number,loff_t *ino,
     struct buffer_head **res_bh,struct msdos_dir_entry **res_de)
 {
 	int count,cluster;
@@ -494,7 +494,7 @@ static int raw_scan_root(struct super_block *sb,const char *name,int *number,int
  */
 
 static int raw_scan_nonroot(struct super_block *sb,int start,const char *name,
-    int *number,int *ino,struct buffer_head **res_bh,struct msdos_dir_entry
+    int *number,loff_t *ino,struct buffer_head **res_bh,struct msdos_dir_entry
     **res_de)
 {
 	int count,cluster;
@@ -530,7 +530,7 @@ static int raw_scan_nonroot(struct super_block *sb,int start,const char *name,
  */
 
 static int raw_scan(struct super_block *sb, int start, const char *name,
-    int *number, int *ino, struct buffer_head **res_bh,
+    int *number, loff_t *ino, struct buffer_head **res_bh,
     struct msdos_dir_entry **res_de)
 {
 	if (start) return raw_scan_nonroot
@@ -621,7 +621,7 @@ int fat_subdirs(struct inode *dir)
  */
 
 int fat_scan(struct inode *dir,const char *name,struct buffer_head **res_bh,
-    struct msdos_dir_entry **res_de,int *ino)
+    struct msdos_dir_entry **res_de,loff_t *ino)
 {
 	int res;
