+++ include/linux/msdos_fs_i.h
@@ -30,7 +30,7 @@ struct msdos_inode_info {
 	int i_attrs;	/* unused attribute bits */
 	int i_ctime_ms;	/* unused change time in milliseconds */
 	int i_binary;	/* file contains non-text data */
-	int i_location;	/* on-disk position of directory entry or 0 */
+	loff_t i_location;	/* on-disk position of directory entry or 0 */
 	struct inode *i_fat_inode;	/* struct inode of this one */
 	struct list_head i_fat_hash;	/* hash by i_location */
 	off_t i_last_pos;/* position of last lookup */
